requirements.txt
flask
scikit-learn
pandas
numpy
nltk
joblib
flask-cors
gunicorn
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
